feat: add Codex limits slash command
